% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusedanova.R
\name{AIC.fusedanova}
\alias{AIC.fusedanova}
\title{AIC of a fusedanova object}
\usage{
\method{AIC}{fusedanova}(object, ngroups = 1:nrow(object$path),
  heights = NULL, k = 2)
}
\description{
compute AIC of a fusedanova object
}
