% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusedanova_S3methods.R
\name{as.hclust.fusedanova}
\alias{as.hclust.fusedanova}
\title{export to hclust format}
\usage{
\method{as.hclust}{fusedanova}(object, ...)
}
\description{
export a fusedanova_cpp output to an hclust object - internal used only
}
