\docType{data}
\name{aves}
\alias{aves}
\title{Aves data set: give birth weights two levels of phylogenetic classifications}
\format{A dataframe with 184 rows with three columns

 \describe{
\item{"weight"}{a numeric vector: weights in gramme.}

\item{"family"}{a factor with 40 levels: the bird family.}

\item{"order"}{a factor with 15 levels: the bird order.}

}}
\source{
  http://genomics.senescence.info/species/
}
\description{
  This data set gives the birth weight of for $40$ bird
  families classified in $15$ orders and regrouping a total
  of $n=184$ individuals.
}
\keyword{datasets}

