% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusedanova_S3methods.R
\name{BIC.fusedanova}
\alias{BIC.fusedanova}
\title{BIC of a fusedanova object}
\usage{
\method{BIC}{fusedanova}(object, ngroups = 1:nrow(object$path),
  heights = NULL)
}
\description{
compute BIC of a fusedanova object
}
