% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusedANOVA-R6Class.R
\docType{data}
\name{fusedANOVA}
\alias{fusedANOVA}
\title{An R6 Class for representing and performing optimization of a fused-ANOVA model}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
fusedANOVA
}
\description{

}
\keyword{datasets}
