% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusedanova_old.R
\name{fusedanova_old}
\alias{fusedanova_old}
\title{Fit a Fused ANOVA model with old version of the package}
\usage{
fusedanova_old(x, class = 1:length(x), weights = c("default", "laplace",
  "gaussian", "adaptive", "personal"), standardize = FALSE, ...)
}
\description{
Adjust a penalized ANOVA model with Fused-LASSO (or Total Variation) penality, 
ie. a sum of weighted \eqn{\ell_1}{l1}-norm on the difference of each coefficient.
}
