% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusedanova_S3methods.R
\name{logLik.fusedanova}
\alias{logLik.fusedanova}
\title{compute loglikelihood of a fusedanova object}
\usage{
\method{logLik}{fusedanova}(object, ngroups = 1:(nlevels(object$group) - 1),
  heights = NULL)
}
\description{
compute loglikelihood of a fusedanova object
}
