\docType{methods}
\name{plot,fusedanova-method}
\alias{plot,fusedanova-method}
\alias{plot.fusedanova}
\title{Plot method for a fusedanova object}
\arguments{
  \item{x}{an object of class \code{fusedanova}.}

  \item{y}{used for S4 compatibility.}

  \item{common}{type of plot. By default \code{"l"}.}

  \item{main}{the main title, with a hopefully appropriate
  default definition.}

  \item{xlab}{character or expression (or a "grob") giving
  label(s) for the x-axis.}

  \item{ylab}{character or expression (or "grob") giving
  label for the y-axis.}

  \item{strip}{logical flag or function. If FALSE, strips
  are not drawn.}

  \item{layout}{the layout of the plot, \code{NULL} by
  default.}

  \item{...}{used for S4 compatibility.}
}
\description{
  Produce a plot of the solution path of a
  \code{fusedanova} fit.
}
\examples{
\dontrun{
## Not done yet
}
}
\seealso{
  \code{\linkS4class{fusedanova}}.
}

