% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ward_1d.R
\name{ward_1d}
\alias{ward_1d}
\title{Fit a fast version of the 1 dimensional Ward clustering algorithm}
\usage{
ward_1d(x, group, hclust = TRUE)
}
\arguments{
\item{x}{a numeric vector of observation for n individuals.}

\item{group}{an optional vector or factor giving the initial grouping. If missing, 
each individual are set in a single group.}

\item{hclust}{boolean: should the result be outputed as an hclust object? Default is \code{TRUE}.}
}
\value{
an S3 object with class \code{hclust} or a data frame of the succesive fusions.
}
\description{
The Ward clustering algorithm is the more statistically grounded hierarchical clustering approach.
This version is dedicated to clustering of univariate data (i.e., vector).
}
\examples{
data(aves)
ward1d <- ward_1d(aves$weight, aves$family)
plot(ward1d)

}
