% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusedanova_multivariate.R,
%   R/fusedanova_univariate.R, R/fusedanova_S3methods.R
\name{fusedanova.data.frame}
\alias{fusedanova.data.frame}
\alias{fusedanova.numeric}
\alias{fusedanova}
\alias{fusedanova.matrix}
\title{Fit a Fused ANOVA model}
\usage{
\method{fusedanova}{data.frame}(x, group, weighting = c("laplace", "gaussian",
  "adaptive"), gamma = rep(0, ncol(x)), standardize = FALSE, W = NULL)

\method{fusedanova}{numeric}(x, group, weighting = c("laplace", "gaussian",
  "adaptive"), gamma = 0, standardize = FALSE, W = NULL, hclust = TRUE)

fusedanova(x, ...)

\method{fusedanova}{matrix}(x, group = 1:nrow(x), weighting = c("laplace",
  "gaussian", "adaptive"), gamma = rep(0, ncol(x)), standardize = TRUE,
  W = NULL)
}
\arguments{
\item{x}{a vector, matrix or data.frame of observation for n individuals.}

\item{group}{vector or factor giving the initial group of each individual. If missing, 
each individual are set in a single group (clustering mode).}

\item{weighting}{character; which type of weights is supposed to be used.
The supported weights are: \code{"laplace"}, \code{"gaussian"} or  \code{"adaptive"}.
See details below. By default, its value is \code{"laplace"}. Ignore if \code{W} is not \code{NULL}}

\item{gamma}{non-negative scalar (or vector if x is a matrix or a data frame);
 the \eqn{\gamma}{gamma} parameter needed for
\code{"laplace"}, \code{"gaussian"} and \code{"adaptive"} weights. Default is 0.}

\item{standardize}{logical; should the vector be standardized before computation?
Default is \code{FALSE}. If \code{TRUE}, the vector is centered and scaled according to the pooled variance of initial groups.}

\item{W}{a numeric matrix of weights of user defined weights. Default is \code{NULL}. 
If not \code{NULL}, should be a k x k matrix (with k the initial number of groups) that
will overwrite the \code{weighting} parameter.}

\item{hclust}{boolean: should the result be outputed as an hclust object? Default is \code{TRUE}.}
}
\value{
an S3 object with class \code{hclust}.

The optimization problem solved by fused-ANOVA is
\if{latex}{\deqn{%
\hat{\beta}_{\lambda} = \arg \min_{\beta}
\left\{\sum_{k=1}^K \sum_{i=1}^{n_k} \left(Y_{ik}-\beta_k \right)^2
+ \lambda \sum_{k,\ell} w_{kl} \left|\beta_k - \beta_\ell \right|\right\}}}
\if{html}{\out{ <center> &beta;<sup>hat</sup>
<sub>&lambda;<sub>1</sub></sub> =
argmin<sub>&beta;</sub> sum<sub>k</sub> sum_i (Y<sub>ik</sub> - &beta<sub>k</sub>)<sup>2</sup>
+ &lambda; sum<sub>k,l</sub> w<sub>k,l</sub>
&#124; &beta;<sub>k</sub> - &beta;<sub>l</sub> &#124;, </center> }}
\if{text}{\deqn{beta.hat(lambda) = argmin_beta sum_k sum_i (Y_ik - beta_k)^2
+ lambda sum_k sum_l w_kl | beta_k - beta_l|,}}

where \eqn{Y_{ik}}{Y_ik} is the intensity of a continuous random
variable for sample \eqn{i}{i} in condition \eqn{k}{k} and
\eqn{\beta_k}{beta_k} is the mean parameter of condition
\eqn{k}{k}. We denote by \eqn{K}{K} the total number of conditions
and \eqn{n_k}{n_k} the number of sample in each condition.
}
\description{
Adjust a penalized ANOVA model with Fused-LASSO (or Total Variation) penality, 
ie. a sum of weighted \eqn{\ell_1}{l1}-norm on the difference of each coefficient.
}
\examples{
data(aves)
fa.laplace <- fusedanova(aves$weight, aves$family, gamma = 0)
plot(fa.laplace)

fa.ada <- fusedanova(aves$weight, aves$family, "adaptive", gamma = 2)
plot(fa.ada)

}
\keyword{models,}
\keyword{regression}
